pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "light weight RAG system that communicates with a SQL database"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.11.0.post1",
    "fastapi>=0.116.1",
    "invoke>=2.2.0",
    "openai>=1.99.5",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "uvicorn>=0.35.0",
]

[tool.uv.workspace]
members = [
    "apps/API/api",
    "apps",
]
